pyproject.toml
[tool.poetry]
name = "ftn-tst"
version = "0.1.0"
description = "test_task_fittin"
authors = ["Eugene Popov <kosmokot45@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
